+++ bare-metal/fedora-coreos/kubernetes/variables.tf
@@ -165,3 +165,15 @@ variable "cluster_domain_suffix" {
   default     = "cluster.local"
 }
 
+variable "components" {
+  description = "Configure pre-installed cluster components"
+  # Component configs are passed through to terraform-render-bootstrap,
+  # which handles type enforcement and defines defaults
+  # https://github.com/poseidon/terraform-render-bootstrap/blob/main/variables.tf#L95
+  type = object({
+    enable     = optional(bool)
+    coredns    = optional(map(any))
+    kube_proxy = optional(map(any))
+  })
+  default = null
+}